backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
python-dotenv
deepgram-sdk
psycopg2-binary
Flask-SocketIO
gevent
gevent-websocket